feat: add confirm dialog delete

+++ src/app/components/sidebar/sidebar.component.ts
@@ -2,6 +2,8 @@ import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {ApiService} from "../../services/api.service";
 import {Router} from "@angular/router";
 import {MatSnackBar} from "@angular/material/snack-bar";
+import {MatDialog} from "@angular/material/dialog";
+import {DialogConfirmComponent} from "../dialog-confirm/dialog-confirm.component";
 
 @Component({
   selector: 'app-sidebar',
@@ -9,7 +11,7 @@ import {MatSnackBar} from "@angular/material/snack-bar";
   styleUrls: ['./sidebar.component.scss']
 })
 export class SidebarComponent implements OnInit{
-  constructor(private api: ApiService, private router: Router, private snack: MatSnackBar) {}
+  constructor(private api: ApiService, private router: Router, private snack: MatSnackBar, private dialog: MatDialog) {}
   panelOpenState = false;
   showFiller = false;
   public list:any = []
@@ -46,14 +48,29 @@ export class SidebarComponent implements OnInit{
   }
 
   deleteItem(items: any) {
-    this.token = localStorage.getItem("key")
-    if(this.token) {
-      this.api.deleteItem(items.id, this.token).subscribe({
-        next: () => {
-          this.snack.open("JSON Removido com sucesso.")
-          this.getAll();
+    const dialog = this.dialog.open(DialogConfirmComponent, {
+      data: {
+        title: 'Excluir',
+        text: 'Tem certeza que deseja excluir este json?'
+      }
+    })
+    dialog.afterClosed().subscribe({
+      next: (resp) => {
+        if(resp) {
+          if(resp) {
+            this.token = localStorage.getItem("key")
+            if(this.token) {
+              this.api.deleteItem(items.id, this.token).subscribe({
+                next: () => {
+                  this.snack.open("JSON Removido com sucesso.")
+                  this.getAll();
+                }
+              })
+            }
+          }
         }
-      })
-    }
+      }
+    })
+
   }
 }
